Move EventLog component
